requirements.txt
pygame>=2.0.0
pillow>=8.0.0
pyinstaller>=5.0.0
